requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.2.10
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.4
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
cohere==3.10.0
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.5.8
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
fastjsonschema==2.17.1
filelock==3.12.2
flake8==6.0.0
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gyp==0.1
h11==0.14.0
h5py==3.9.0
httpcore==0.17.2
httplib2==0.20.2
httpx==0.25.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
jax==0.4.12
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
keyring==23.5.0
langchain==0.0.208
langchainplus-sdk==0.0.16
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lit==16.0.6
manifest-ml==0.0.1
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.1
ml-dtypes==0.2.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
nlpcloud==1.0.42
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
openlm==0.0.5
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymongo==4.3.3
pyparsing==2.4.7
pyrsistent==0.19.3
pyTelegramBotAPI==4.12.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-telegram-bot==20.6
pytube==15.0.0
PyYAML==5.4.1
pyzmq==25.1.0
redis==4.5.5
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlitedict==2.1.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
systemd-python==234
telebot==0.0.5
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.6.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.16
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
Werkzeug==2.3.6
wrapt==1.14.1
yarl==1.9.2
zipp==1.0.0
